+++ zeroize/src/aarch64.rs
@@ -1,6 +1,6 @@
 //! [`Zeroize`] impls for ARM64 SIMD registers.
 
-use crate::{atomic_fence, volatile_write, Zeroize};
+use crate::{Zeroize, atomic_fence, volatile_write};
 
 use core::arch::aarch64::*;