+++ block-buffer/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ReadBuffer` type ([#823])
 - Optional implementation of the `Zeroize` trait ([#963])
 - Generic `serialize` and `deserialize` methods ([#1200])
+- `ReadBuffer::{read_cached, write_block}` methods ([#1201])
 
 ### Changed
 - Block sizes must be bigger than 0 and smaller than 256.
@@ -19,12 +20,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 - `EagerBuffer::set_data` method. Use the `ReadBuffer` type instead. ([#823])
+- `ReadBuffer::read` method ([#1201])
 
 [#963]: https://github.com/RustCrypto/utils/pull/963
 [#1115]: https://github.com/RustCrypto/utils/pull/1115
 [#1115]: https://github.com/RustCrypto/utils/pull/1116
 [#1149]: https://github.com/RustCrypto/utils/pull/1149
 [#1200]: https://github.com/RustCrypto/utils/pull/1200
+[#1201]: https://github.com/RustCrypto/utils/pull/1201
 
 ## 0.10.3 (2022-09-04)
 ### Added