+++ block-buffer/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## UNRELEASED
+## 0.11.0 (2025-11-07)
 ### Added
 - `ReadBuffer` type ([#823])
 - Optional implementation of the `Zeroize` trait ([#963])